Null checking and less casts
